Refactor dark mode toggle and update theme handling.
Updated App.css to use daisyui plugin with data-theme attribute for dark mode.

Updated DarkModeToggle.razor to use a fieldset with a checkbox for toggling dark mode and directly invoke the JavaScript function to set the theme.

Modified DarkModeToggle.razor.js to update the data-theme attribute on the main-layout element. Added an id attribute to the main div in MainLayout.razor.

Updated Home.razor with a new h1 element and adjusted margin-top for buttons in dark mode.

Uncommented inspectUri settings in launchSettings.json for browser debugging.